[r] Move all equal's helpers into class Equal
- As all currents helpers only serve class Equal, let's move them
inside Equal
