+++ robber/matchers/equal.py
@@ -2,7 +2,6 @@ from difflib import Differ
 
 from robber import expect
 from robber.explanation import Explanation
-from robber.helper import unicode_to_str
 from robber.matchers.base import Base
 
 
@@ -67,8 +66,46 @@ B['{key}'] = {val_b}
         return 'Diffs:\n{0}'.format('\n'.join(diffs))
 
     def standardize_args(self):
-        self.actual = unicode_to_str(self.actual)
-        self.expected = unicode_to_str(self.expected)
+        self.actual = self.unicode_to_str(self.actual)
+        self.expected = self.unicode_to_str(self.expected)
+
+    @classmethod
+    def unicode_to_str(cls, obj):
+        try:
+            if type(obj) is unicode:
+                return cls._unicode_string_to_str(obj)
+        except NameError:
+            pass
+
+        if type(obj) is list:
+            return cls._unicode_list_to_str_list(obj)
+
+        if type(obj) is dict:
+            return cls._unicode_dict_to_str_dict(obj)
+
+        return obj
+
+    @classmethod
+    def _unicode_string_to_str(cls, u_string):
+        return u_string.encode('utf-8')
+
+    @classmethod
+    def _unicode_list_to_str_list(cls, u_list):
+        str_list = list(u_list)
+
+        for i in range(0, len(str_list)):
+            str_list[i] = cls.unicode_to_str(str_list[i])
+
+        return str_list
+
+    @classmethod
+    def _unicode_dict_to_str_dict(cls, u_dict):
+        str_dict = dict(u_dict)
+
+        for key in str_dict:
+            str_dict[key] = cls.unicode_to_str(str_dict[key])
+
+        return str_dict
 
 
 expect.register('eq', Equal)
